Consolidated migrations - as this currently isn't live then there isn't a problem.  Made rating a Double to allow the same model to be used for rating and average rating.

+++ Open.GI.hypermart/Models/Rating.cs
@@ -53,8 +53,7 @@ namespace Open.GI.hypermart.Models
         /// The rating.
         /// </value>
         /// 
-        //TODO : This should be a double.
-        public int rating { get; set; }
+        public double rating { get; set; }
 
         /// <summary>
         /// Gets or sets the comments.
